+++ sdk/resourcemanager/containerservice/armcontainerservice/constants.go
@@ -7,7 +7,7 @@ package armcontainerservice
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
-	moduleVersion = "v7.0.0-beta.2"
+	moduleVersion = "v7.0.0-beta.3"
 )
 
 // AddonAutoscaling - Whether VPA add-on is enabled and configured to scale AKS-managed add-ons.
@@ -350,6 +350,24 @@ func PossibleFormatValues() []Format {
 	return []Format{
 		FormatAzure,
 		FormatExec,
+	}
+}
+
+// GPUDriver - Whether to install GPU drivers. When it's not specified, default is Install.
+type GPUDriver string
+
+const (
+	// GPUDriverInstall - Install driver.
+	GPUDriverInstall GPUDriver = "Install"
+	// GPUDriverNone - Skip driver install.
+	GPUDriverNone GPUDriver = "None"
+)
+
+// PossibleGPUDriverValues returns the possible values for the GPUDriver const type.
+func PossibleGPUDriverValues() []GPUDriver {
+	return []GPUDriver{
+		GPUDriverInstall,
+		GPUDriverNone,
 	}
 }
 
@@ -840,6 +858,29 @@ func PossibleNodeOSUpgradeChannelValues() []NodeOSUpgradeChannel {
 		NodeOSUpgradeChannelNone,
 		NodeOSUpgradeChannelSecurityPatch,
 		NodeOSUpgradeChannelUnmanaged,
+	}
+}
+
+// NodeProvisioningDefaultNodePools - This field has no effect unless mode is 'Auto'. Warning: Changing this from Auto to
+// None on an existing cluster will cause the default Karpenter NodePools to be deleted, which will in turn drain and
+// delete the nodes associated with those pools. It is strongly recommended to not do this unless there are idle nodes ready
+// to take the pods evicted by that action. If not specified, the default is
+// Auto. For more information see aka.ms/something
+type NodeProvisioningDefaultNodePools string
+
+const (
+	// NodeProvisioningDefaultNodePoolsAuto - A standard set of Karpenter NodePools are provisioned
+	NodeProvisioningDefaultNodePoolsAuto NodeProvisioningDefaultNodePools = "Auto"
+	// NodeProvisioningDefaultNodePoolsNone - No Karpenter NodePools are provisioned automatically. Automatic scaling will not
+	// happen unless the user creates one or more NodePool instances.
+	NodeProvisioningDefaultNodePoolsNone NodeProvisioningDefaultNodePools = "None"
+)
+
+// PossibleNodeProvisioningDefaultNodePoolsValues returns the possible values for the NodeProvisioningDefaultNodePools const type.
+func PossibleNodeProvisioningDefaultNodePoolsValues() []NodeProvisioningDefaultNodePools {
+	return []NodeProvisioningDefaultNodePools{
+		NodeProvisioningDefaultNodePoolsAuto,
+		NodeProvisioningDefaultNodePoolsNone,
 	}
 }